fix(trainable): remove `dtype` enforcement
Filter parameters were enforced to `torch.float32`. This is removed.

+++ trainable/filtering.py
@@ -16,7 +16,7 @@ class TrainableDiagonalFilter(nn.Module):
         super().__init__()
         self.size = size
         self.device = None
-        self.filter = nn.Parameter(torch.randn(size, dtype=torch.float32))
+        self.filter = nn.Parameter(torch.randn(size))
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         if self.device is None:
